+++ backend/internal/students/domain.go
@@ -227,8 +227,8 @@ type ImportRequest struct {
 // HTTP status should be 200 instead of 201.
 type ImportResponse struct {
 	JobID        string `json:"job_id"`
-	TotalRecords int    `json:"total_records"`
-	TotalChunks  int    `json:"total_chunks"`
+	TotalRecords int64  `json:"total_records"`
+	TotalChunks  int64  `json:"total_chunks"`
 	Status       string `json:"status"`
 	IsReplay     bool   `json:"is_replay"`
 }
